move pick_building member
