+++ src/compas_fab/backends/vrep/client.py
@@ -284,32 +284,6 @@ class VrepClient(ClientInterface):
 
         return final_states
 
-    def pick_building_member(self, robot, building_member_mesh, pickup_frame, metric_values=None):
-        """Picks up a building member and attaches it to the robot.
-
-        Args:
-            robot (:class:`compas_fab.robots.Robot`): Robot instance to use for pick up.
-            building_member_mesh (:class:`compas.datastructures.Mesh`): Mesh
-                of the building member that will be attached to the robot.
-            pickup_frame (:class:`Frame`): Pickup frame.
-            metric_values (:obj:`list` of :obj:`float`): List containing one value
-                per configurable joint. Each value ranges from 0 to 1,
-                where 1 indicates the axis/joint is blocked and cannot
-                move during inverse kinematic solving.
-
-        Returns:
-            int: Object handle (identifier) assigned to the building member.
-        """
-        assert_robot(robot)
-
-        joints = len(robot.get_configurable_joints())
-        if not metric_values:
-            metric_values = [0.1] * joints
-
-        self.set_robot_pose(robot, pickup_frame)
-
-        return self.add_attached_collision_mesh(building_member_mesh, options={'robot_name': robot.name})
-
     def add_meshes(self, meshes):
         """Adds meshes to the 3D scene.
